requirements.txt
absl-py==2.1.0
accelerate==1.1.0
aiohttp==3.8.5
aiosignal==1.3.1
alembic==1.11.2
annotated-types==0.7.0
anyio==4.6.2.post1
aria==0.4.5
aria2==0.0.1b0
astunparse==1.6.3
async-timeout==4.0.3
attrs==23.1.0
azure-common==1.1.28
azure-core==1.29.2
azure-identity==1.14.0
azure-mgmt-core==1.4.0
azure-storage-blob==12.13.0
azureml-mlflow==1.52.0
beautifulsoup4==4.12.3
bert-score==0.3.5
bitsandbytes==0.45.0
blinker==1.6.2
bypy==1.8.5
cachetools==5.5.0
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.6
cloudpickle==2.2.1
compressed-tensors==0.6.0
contourpy==1.1.0
cryptography==41.0.3
cycler==0.11.0
dacite==1.8.1
databricks-cli==0.17.7
databricks-sdk==0.36.0
datasets==3.2.0
deepspeed==0.16.0
Deprecated==1.2.14
dill==0.3.7
diskcache==5.6.3
distro==1.9.0
docker==6.1.3
docker-pycreds==0.4.0
einops==0.8.0
exceptiongroup==1.1.3
fairscale==0.4.13
faiss-cpu==1.9.0
fastapi==0.115.4
fastchat==0.1.0
filelock==3.16.1
fire==0.7.0
flash_attn==2.7.0.post2
Flask==2.3.2
flatbuffers==24.3.25
fonttools==4.42.0
frozenlist==1.4.0
fsspec==2023.6.0
functorch==0.2.1
gast==0.6.0
gdown==5.2.0
gguf==0.10.0
gitdb==4.0.10
GitPython==3.1.32
google-auth==2.35.0
google-pasta==0.2.0
GPUtil==1.4.0
graphene==3.4.1
graphql-core==3.2.5
graphql-relay==3.2.0
graphviz==0.20.3
greenlet==2.0.2
grpcio==1.67.1
gunicorn==21.2.0
h11==0.14.0
h5py==3.12.1
hjson==3.1.0
htmlmin==0.1.12
httpcore==1.0.6
httptools==0.6.4
httpx==0.27.2
huggingface-hub==0.26.2
idna==3.4
ImageHash==4.3.1
imageio==2.36.0
importlib-metadata==6.8.0
iniconfig==2.0.0
interegular==0.3.3
isodate==0.6.1
itsdangerous==2.1.2
jiter==0.6.1
joblib==1.3.2
jsonpatch==1.33
jsonpickle==3.0.2
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
keras==3.6.0
kiwisolver==1.4.4
labml==0.4.168
labml-helpers==0.4.89
labml-nn==0.4.137
langchain==0.3.6
langchain-core==0.3.15
langchain-text-splitters==0.3.1
langsmith==0.1.139
lark==1.2.2
libclang==18.1.1
litellm==1.59.9
llvmlite==0.43.0
lm-format-enforcer==0.10.6
loralib==0.1.2
Mako==1.2.4
Markdown==3.4.4
markdown-it-py==3.0.0
matplotlib==3.7.2
mauve-text==0.4.0
mdurl==0.1.2
metrics==0.3.3
mistral_common==1.4.4
ml-dtypes==0.4.1
ml-swissknife==0.1.30
mlflow==2.17.2
mlflow-skinny==2.17.2
msal==1.23.0
msal-extensions==1.0.0
msgpack==1.1.0
msgspec==0.18.6
msrest==0.7.1
multidict==6.0.4
multimethod==1.12
multiprocess==0.70.15
namex==0.0.8
ninja==1.11.1.1
nltk==3.5
numba==0.60.0
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-ml-py==12.535.161
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.1.105
nvitop==1.3.2
oauthlib==3.2.2
opacus==1.5.2
openai==1.60.2
opencv-python-headless==4.10.0.84
opentelemetry-api==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-semantic-conventions==0.48b0
opt-einsum==3.3.0
optree==0.13.0
orjson==3.10.10
outlines==0.0.46
pandas==2.0.3
partial-json-parser==0.2.1.1.post4
pathlib2==2.3.7.post1
pathspec==0.5.5
patsy==0.5.6
peft==0.13.2
phik==0.12.4
pillow==10.4.0
pluggy==1.2.0
portalocker==2.7.0
progress==1.6
prometheus-fastapi-instrumentator==7.0.0
prometheus_client==0.21.0
protobuf==5.28.3
prv-accountant==0.1.1.post1
py-cpuinfo==9.0.0
pyairports==2.1.1
pyarrow==19.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycountry==24.6.1
pycparser==2.21
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
PyJWT==2.8.0
pyparsing==3.0.9
PySocks==1.7.1
pyter3==0.3
pytest==7.4.0
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2023.3
PyWavelets==1.7.0
PyYAML==6.0.1
querystring-parser==1.2.4
ray==2.38.0
razdel==0.5.0
referencing==0.35.1
regex==2023.8.8
requests==2.32.3
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
responses==0.18.0
rich==13.9.4
rouge-chinese==1.0.3
rouge-score==0.1.2
rpds-py==0.21.0
rsa==4.9
safetensors==0.4.5
scikit-learn==1.3.0
scipy==1.11.1
seaborn==0.13.2
sentence-transformers==3.4.1
sentencepiece==0.2.0
sentry-sdk==2.17.0
setproctitle==1.3.3
smmap==5.0.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.19
sqlparse==0.4.4
starlette==0.41.2
statsmodels==0.14.4
sympy==1.13.1
tabulate==0.8.7
tenacity==9.0.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.5.0
textgrad==0.1.6
tf-slim==1.1.0
threadpoolctl==3.2.0
tiktoken==0.7.0
tokenizers==0.21.0
tomli==2.0.1
torch==2.4.1
torchaudio==2.4.1
torchtext==0.18.0
torchvision==0.19.1
tqdm==4.67.1
transformers==4.47.0
triton==3.0.0
typeguard==4.4.0
typing_extensions==4.12.2
tzdata==2023.3
urllib3==1.26.16
uvicorn==0.32.0
uvloop==0.21.0
visions==0.7.6
vllm==0.6.3.post1
wandb==0.18.5
watchfiles==0.24.0
websocket-client==1.6.1
websockets==14.0
Werkzeug==2.3.7
wordcloud==1.9.3
wrapt==1.16.0
xformers==0.0.27.post2
xxhash==3.3.0
yarl==1.9.2
ydata-profiling==4.12.0
zipp==3.16.2
